+++ src/importlinter/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.11"
+__version__ = "2.12"
 
 from .application import output  # noqa
 from .domain import fields  # noqa