+++ docs/release_notes.md
@@ -3,6 +3,7 @@
 ## latest
 
 * Improve error message when root package is a single-file module.
+* Alert users with all unmatched ignored imports in the same run.
 
 ## 2.11 (2026-03-06)